+++ Src/ImplicitNullability.Plugin/Highlighting/ImplicitNotNullOverridesUnknownBaseMemberNullabilityHighlighting.cs
@@ -4,15 +4,16 @@ using JetBrains.ReSharper.Psi.Tree;
 using ReSharperExtensionsShared.Highlighting;
 
 [assembly: RegisterConfigurableSeverity(
-    ImplicitNotNullOverridesUnknownExternalMemberHighlighting.SeverityId,
+    ImplicitNotNullOverridesUnknownBaseMemberNullabilityHighlighting.SeverityId,
     null,
     HighlightingGroupIds.CodeSmell,
-    ImplicitNotNullOverridesUnknownExternalMemberHighlighting.Message,
-    ImplicitNotNullOverridesUnknownExternalMemberHighlighting.Description,
+    ImplicitNotNullOverridesUnknownBaseMemberNullabilityHighlighting.Message,
+    ImplicitNotNullOverridesUnknownBaseMemberNullabilityHighlighting.Description,
     Severity.WARNING
 #if RESHARPER20162
     , SolutionAnalysisRequired: false
 #endif
+    , AlternativeIDs = "ImplicitNotNullOverridesUnknownExternalMember"
 )]
 
 namespace ImplicitNullability.Plugin.Highlighting
@@ -22,24 +23,22 @@ namespace ImplicitNullability.Plugin.Highlighting
         "CSHARP",
         OverlapResolve = OverlapResolveKind.WARNING,
         ToolTipFormatString = Message)]
-    public class ImplicitNotNullOverridesUnknownExternalMemberHighlighting : SimpleTreeNodeHighlightingBase<ITreeNode>
+    public class ImplicitNotNullOverridesUnknownBaseMemberNullabilityHighlighting : SimpleTreeNodeHighlightingBase<ITreeNode>
     {
-        // IDEA: Due to symmetry reasons with ImplicitNotNullConflictInHierarchyHighlighting and to emphasize that this is about
-        // input/ref elements, rename to ImplicitNotNullPotentialConflictWithExternalBaseMember?
+        public const string SeverityId = "ImplicitNotNullOverridesUnknownBaseMemberNullability";
 
-        public const string SeverityId = "ImplicitNotNullOverridesUnknownExternalMember";
-        public const string Message = "Implicit NotNull overrides unknown nullability of external code, nullability should be explicit";
+        public const string Message = "Implicit NotNull overrides unknown nullability of base member, nullability should be explicit";
 
         public const string Description =
-            "Warns about implicit [NotNull] elements that override an external base class/interface member " +
+            "Warns about implicit [NotNull] elements that override a base class/interface member " +
             "which has no corresponding nullability annotations (neither by attributes nor by external XML annotations). " +
             "Because the base member's nullability is unknown, we do not know whether a substitutability " +
             "violation exists (e.g. implicit [NotNull] parameter overrides base member with an unannotated " +
-            "parameter, which has [CanBeNull] semantic), we thus encourage the programmer to explicitly " +
+            "parameter which actually has [CanBeNull] semantics), we thus encourage the programmer to explicitly " +
             "annotate [NotNull] or [CanBeNull] on those elements after manually checking the nullability of the base member. " +
             SharedHighlightingTexts.NeedsSettingNoteText;
 
-        public ImplicitNotNullOverridesUnknownExternalMemberHighlighting(ITreeNode treeNode)
+        public ImplicitNotNullOverridesUnknownBaseMemberNullabilityHighlighting(ITreeNode treeNode)
             : base(treeNode, Message)
         {
         }